renamed compoent to match React component naming standards.  Moved shared component into its own folder
