+++ client/src/components/Reviews/ReviewTile.jsx
@@ -1,7 +1,7 @@
 import React, {useState, useEffect} from "react";
 import axios from 'axios';
 import { format, parseISO } from 'date-fns';
-import { StarRating } from '../shared/starRating.jsx';
+import { StarRating } from '../shared/StarRating/StarRating.jsx';
 
 export const ReviewTile = (props) => {
   const { review } = props;
@@ -9,14 +9,10 @@ export const ReviewTile = (props) => {
 
   return (
     <div>
-      {/* refactor for out of 5 stars with 1/4 increments */}
-      {/* star rating: {review.rating} */}
       <StarRating rating={review.rating} />
       <br/>
-      {/* date of review Month DD, YYYY */}
       {format(parseISO(review.date), 'MMMM dd, yyyy')}
       <br/>
-      {/* Review Summary in BOLD */}
       <b>{review.summary.slice(0,60)}</b>
       <br/>
       {/* Review Body with text and images*/}
